requirements.txt
flask==3.1.0
spotipy==2.24.0
gunicorn==23.0.0
python-dotenv==1.0.1
